prototype/apps/backend/onboarding/src/logic/upload_resume/requirements.txt
functions-framework
firebase-admin
google-cloud-storage